:recycle: Refactor runtime code into its own folder.
