--- src/adopt_stylesheets.mjs
@@ -1,49 +1,0 @@
-
-const copiedStyleSheets = new WeakMap();
-
-export default async function adoptStylesheets(shadowRoot) {
-  const pendingParentStylesheets = [];
-  for (const node of document.querySelectorAll("link[rel=stylesheet], style")) {
-    if (node.sheet) continue;
-
-    pendingParentStylesheets.push(
-      new Promise((resolve, reject) => {
-        node.addEventListener("load", resolve);
-        node.addEventListener("error", reject);
-      }),
-    );
-  }
-
-  await Promise.allSettled(pendingParentStylesheets);
-
-  shadowRoot.adoptedStyleSheets =
-    shadowRoot.host.getRootNode().adoptedStyleSheets;
-
-  const pending = [];
-
-  for (const sheet of document.styleSheets) {
-    try {
-      shadowRoot.adoptedStyleSheets.push(sheet);
-    } catch {
-      try {
-        let copiedSheet = copiedStyleSheets.get(sheet);
-        if (!copiedSheet) {
-          copiedSheet = new CSSStyleSheet();
-          for (const rule of sheet.cssRules) {
-            copiedSheet.insertRule(rule.cssText, copiedSheet.cssRules.length);
-          }
-          copiedStyleSheets.set(sheet, copiedSheet);
-        }
-
-        shadowRoot.adoptedStyleSheets.push(copiedSheet);
-      } catch {
-        const node = sheet.ownerNode.cloneNode();
-
-        shadowRoot.prepend(node);
-        pending.push(node);
-      }
-    }
-  }
-
-  return pending;
-}
