--- src/lustre_escape_ffi.erl
@@ -1,4 +1,0 @@
--module(lustre_escape_ffi).
--export([coerce/1]).
-
-coerce(X) -> X.
